+++ tangsquad/src/main/java/backend/tangsquad/logbook/controller/LogbookController.java
@@ -1,11 +1,9 @@
 package backend.tangsquad.logbook.controller;
 
 import backend.tangsquad.auth.jwt.UserDetailsImpl;
-import backend.tangsquad.like.dto.request.LikeLogbookRequest;
 import backend.tangsquad.like.service.LikeLogbookService;
 import backend.tangsquad.logbook.dto.request.LogbookCreateRequest;
 import backend.tangsquad.logbook.dto.request.LogbookRequest;
-import backend.tangsquad.logbook.dto.response.LogResponse;
 import backend.tangsquad.logbook.dto.response.LogbookResponse;
 import backend.tangsquad.logbook.entity.Logbook;
 import backend.tangsquad.logbook.service.LogbookService;
@@ -18,7 +16,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import retrofit2.http.Path;
 
 import java.util.List;
 
@@ -33,6 +30,14 @@ public class LogbookController {
     private final LogbookService logbookService;
     private final LikeLogbookService likeLogbookService;
 
+    private ResponseEntity<LogbookResponse> checkUserDetailsAndRespond(UserDetailsImpl userDetails) {
+        if (userDetails == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(null);
+        }
+        return null;
+    }
+
     @PostMapping
     @Operation(
             summary = "로그북 생성",
@@ -43,18 +48,24 @@ public class LogbookController {
             @RequestBody LogbookCreateRequest logbookCreateRequest,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
 
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return response;
+
         LogbookResponse logbookResponse = logbookService.save(logbookCreateRequest, userDetails);
 
         if (logbookResponse != null) {
             return ResponseEntity.status(HttpStatus.CREATED).body(logbookResponse);
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
     }
 
     @GetMapping("{logbookId}")
     @Operation(summary = "내 로그북 불러오기", description = "내 로그를 불러옵니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<LogbookResponse> getMyLogbook(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable("logbookId") Long logbookId) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return response;
+
         Logbook logbook = logbookService.getLogbookByIdAndUserId(logbookId, userDetails.getId());
 
         if (logbook != null) {
@@ -68,37 +79,30 @@ public class LogbookController {
     @GetMapping("all")
     @Operation(summary = "모든 로그북 불러오기", description = "모든 로그북을 불러옵니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<List<LogbookResponse>> getAllLogbooks(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
         List<LogbookResponse> logbookResponses = logbookService.getAllLogbooks(userDetails);
 
-        if (logbookResponses != null) {
+        if (logbookResponses != null && !logbookResponses.isEmpty()) {
             return ResponseEntity.ok(logbookResponses);
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
-
     }
 
     @GetMapping("")
     @Operation(summary = "내 로그북 불러오기", description = "나의 로그들을 불러옵니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<List<LogbookResponse>> getMyLogbooks(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
         List<LogbookResponse> logbookResponses = logbookService.getLogbooksByUserId(userDetails);
 
-        if (logbookResponses != null) {
+        if (logbookResponses != null && !logbookResponses.isEmpty()) {
             return ResponseEntity.ok(logbookResponses);
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
-        }
-    }
-
-    @GetMapping("user/{userId}")
-    @Operation(summary = "유저 로그북 불러오기", description = "해당 유저의 로그들을 불러옵니다.")
-    public ResponseEntity<List<LogbookResponse>> getLogbooks(@PathVariable("userId") Long userId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        List<LogbookResponse> logbookResponses = logbookService.getLogbooksByUserId(userDetails);
-
-        if (logbookResponses != null) {
-            return ResponseEntity.ok(logbookResponses);
-        } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
     }
 
@@ -107,21 +111,27 @@ public class LogbookController {
     public ResponseEntity<LogbookResponse> updateLog(
             @RequestBody LogbookRequest logbookRequest,
             @AuthenticationPrincipal UserDetailsImpl userDetailsImpl) {
+
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetailsImpl);
+        if (response != null) return response;
+
         LogbookResponse logbookResponse = logbookService.updateLogbook(logbookRequest, userDetailsImpl);
 
         if (logbookResponse == null) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         } else {
             return ResponseEntity.ok(logbookResponse);
         }
     }
 
-    // Use @PathVariable for the ID since it's in the URL path
     @DeleteMapping("{logbookId}")
     @Operation(summary = "로그북 삭제하기", description = "나의 로그북을 삭제합니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<CommonResponse> deleteLogbook(
             @PathVariable("logbookId") Long logbookId,
             @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 
         return logbookService.deleteLog(logbookId, userDetails);
     }
@@ -132,11 +142,12 @@ public class LogbookController {
         logbookService.deleteAll();
     }
 
-
-    // 로그북 인스턴스 모두 나와야 함.
     @PostMapping("like/{logbookId}")
     @Operation(summary = "좋아요 로그북 추가", description = "로그북에 좋아요를 추가합니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<LogbookResponse> likeLogbook(@PathVariable Long logbookId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return response;
+
         LogbookResponse logbookResponse = likeLogbookService.createLike(logbookId, userDetails);
         return ResponseEntity.ok(logbookResponse);
     }
@@ -144,23 +155,29 @@ public class LogbookController {
     @GetMapping("like")
     @Operation(summary = "좋아요한 로그북 가져오기", description = "좋아요한 로그북을 가져옵니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<List<LogbookResponse>> getLikeLogbooks(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
         List<LogbookResponse> logbookResponses = likeLogbookService.getLikeLogbooks(userDetails);
 
-        if (logbookResponses != null) {
+        if (logbookResponses != null && !logbookResponses.isEmpty()) {
             return ResponseEntity.ok(logbookResponses);
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         }
     }
 
     @DeleteMapping("like/{logbookId}")
     @Operation(summary = "좋아요한 로그북 취소하기", description = "좋아요한 로그북을 취소합니다.", security = @SecurityRequirement(name = "AccessToken"))
     public ResponseEntity<String> cancelLikeLogbooks(@PathVariable("logbookId") Long logbookId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        ResponseEntity<LogbookResponse> response = checkUserDetailsAndRespond(userDetails);
+        if (response != null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
         try {
             likeLogbookService.cancelLike(logbookId, userDetails);
+            return ResponseEntity.ok("좋아요가 취소되었습니다.");
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("좋아요 취소에 실패했습니다.");
         }
-        return ResponseEntity.ok().body("좋아요 취소 완료");
     }
 }
